Rename(#30-battery-view): BatteryChartView 파일 분리
